Added barcode scanner bounding box
Implemented a barcode scanner overlay and bounding box.
Only the first barcode detected inside the bounding box
is passed to the viewModel.
To match the location of the barcode in the preview and imageProxy,
the scale factor between the two the was retrieved and
used to find the equivalent points of the scanner and barcode bounding boxes.

* Add BarcodeAnalyzer.kit file
* Update BarcodeScanner composable to change on valid barcode.
* Update ScannerOverlay instance in BarcodeScanner composable
* Add hasFoundBarcode value in StockUpUiState