Added Functions to Stock Up Screen
Scanner is now either fetches product details from database
and displays information in StockUpBottomSheet, or shows
a dialog asking user to create a new product entry if barcode
has no existing record in the database. When confirmed, the user
is directed to the Product Entry Screen.

Search bar functions similarly. Clicking on a search result
display the bottom sheet with the product detail selected.
An "Add New Product" button directs the user to the Product
Entry screen when pressed.

NOTE: Search bar results do not update with the latest information
unless recomposition occurs.

* Add functions stated above.
* Improve UI design of search bar results list.
* Fix bug where the cursor in search bar is not always at the end of the string.
* Restructured ScannerState and SearchBarState.