Added Receipt Scanner
Started making the barcode scanner for the receipt.
Created two viewmodel, Receipt and ReceiptScanner.
The Receipt viewmodel can be shared between the receipt
and scanner screen, which means we can add products to the
receipt while in the scanner.

Tried using the SavedStateHandle object to pass products
from the scanner to the receipt, but using shared viewmodels
was easier to manage.

* Refactor scanner related classes' names.
* Change ProductDao "getProducts" function result type to nullable.
* Move getProduct function from children to parent BarcodeScannerViewModel
* Add onPermissionResult, SwitchCamera, and onBarcodeScanned functions to
  ScannerCameraState class.