Improved receipt scanner
The ReceiptScannerScreen has been improved with better structure
of composable code and its own bottom sheet similar
to StockUpBottomSheet.

ReceiptScannerViewModel and ReceiptViewModel have been made.
ReceiptViewModel is shared between both it's and the scanner
screen to add products to the receipt's list.

ReceiptViewModel now inserts models into the respective tables
and updates the inventory stocks when sale is confirmed.

* Create ReceiptScannerBottomSheet.
* Add validateInput function in both scanner viewmodels.
* Add utils/Constants.kt file.
* Add SortingType enum.