Update scanner and product entry screens
This commit nests the constraint layouts found in the product
details and entry screens to allow the content to be scrollable
when elements to do not fit. This also changes the content and
element margins of the mentioned screens.

This commit also updates the top app bars found in the receipt and
stock up scanner screens. The top app bar is now a TopAppBar composable
and is transparent. The app bar contents change color depending if the app
has permission to use the camera or not.

This commit also makes the scanner screens rotatable since locking it in
portrait orientation causes the app to crash when navigating back to it
while in landscape.

* Add a BarcodeScannerTopAppBar composable
* Fix issue in product entry and detail screen where
   it is not scrollable on small screen sizes.
* Fix issue in product detail screen where pressing the up
   button navigates back instead of toggling off editing mode first.
* Add message in scanner screen when the app has no camera permissions.
* Fix issues in scanner screens where navigating back to it while in landscape
   orientation crashes the app.