Create Receipt Screen
This commit creates the Receipt Screen along with functions,
Ui, and resources required. This commit adds function such
as removing products from the receipt, clearing receipt,
and confirming sale. This also adds confirmation dialogs and a
modal bottom sheet for the screen.

This commit also refactors some values and resources, and
changes how viewModels are instantiated in each fragment.

* Create Receipt Screen functions and layout.
* Create ReceiptModalBottomSheetFragment functions and layout.
* Create ReceiptListAdapter function and layout.
* Change viewModel initialization in scanner and receipt screens.
* Change confirmSale() in ReceiptViewModel to suspend function
* Change navigation destination IDs and action names.
* Change top and bottom app bar dimens to app_bar_height_{size} format.